Afegits els jocs, cambiat el tipus de resultat

+++ ApiSostenibilitatDef/Models/Game.cs
@@ -18,12 +18,12 @@ namespace ApiSostenibilitat.Models
         [Required]
         [Range(0, 1000000,
         ErrorMessage = "Value for MinRes must be greater than 0.")]
-        public double MinRes { get; set; }
+        public int MinRes { get; set; }
 
         [Required]
         [Range(0.1, 999999,
         ErrorMessage = "Value for MaxRes must be greater than 0.1")]
-        public double MaxRes { get; set; }
+        public int MaxRes { get; set; }
         [Required]
         public string Name {  get; set; }
